refactor: move validate testcase code from validator to loader
