+++ httprunner/cli.py
@@ -5,11 +5,11 @@ import sys
 from httprunner import __description__, __version__
 from httprunner.api import HttpRunner
 from httprunner.compat import is_py2
+from httprunner.loader import validate_json_file
 from httprunner.logger import color_print
 from httprunner.report import gen_html_report
 from httprunner.utils import (create_scaffold, get_python2_retire_msg,
                               prettify_json_file)
-from httprunner.validator import validate_json_file
 
 
 def main():
